Split up chalk module a bit
